src/requirements.txt
mysql-connector-python==8.0.28
openpyxl==3.0.9
selenium==4.1.0